peacock_model/peacock_requirements.txt
# Requirements for Peacock Arabic-English Image Captioning Experiment
# Install with: pip install -r peacock_requirements.txt

# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Image processing
Pillow>=9.0.0
requests>=2.28.0

# Experiment tracking and logging
wandb>=0.15.0
python-dotenv>=1.0.0

# Data handling
pandas>=1.5.0
numpy>=1.24.0

# Optional: For better performance with CUDA
# accelerate>=0.20.0

# Note: Make sure you have CUDA installed if you want to use GPU acceleration
# You can check CUDA availability with: python -c "import torch; print(torch.cuda.is_available())"
